profile info

+++ src/app/components/profile/profile.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from 'src/app/services/auth.service';
 import { Router } from '@angular/router';
-import { timeStamp } from 'console';
 
 @Component({
   selector: 'app-profile',
@@ -9,28 +8,39 @@ import { timeStamp } from 'console';
   styleUrls: ['./profile.component.less']
 })
 export class ProfileComponent implements OnInit {
-  constructor(private auth: AuthService, private router: Router) { }
+  email: string;
+  name: string;
+  number: string;
 
-  ngOnInit(): void {
-    this.auth.GetUser().subscribe(
-      user => {
-        document.getElementById('name').innerHTML = user.username
-      },
-      error => {
-        this.router.navigateByUrl('login')
-      }
-    )
-  }
+	constructor(private auth: AuthService, private router: Router) { }
 
-  deleteProfile() {
-    this.auth.DeleteUser().subscribe(
-      res => {
-        this.router.navigateByUrl('login')
-      }, 
-      err => {
-        alert(err.error)
-      }
-    )
-  }
+	ngOnInit(): void {
+		this.auth.GetUser().subscribe(
+			user => {
+				let num = user.phone_number
+				this.email = user.email
+				this.name = user.username
+				this.number = num.slice(0, 3) + '-' + num.slice(3, 6) + '-' + num.slice(6)
+			},
+			error => {
+				this.router.navigateByUrl('login')
+			}
+		)
+	}
+
+	deleteProfile() {
+		let del = confirm('Are you sure you want to delete your profile? This action is unreversible');
+		if (del) {
+			this.auth.DeleteUser().subscribe(
+				res => {
+					this.router.navigateByUrl('login')
+				}, 
+				err => {
+					alert(err.error)
+				}
+			)
+		}
+
+	}
 
 }
